refactor: panel geometry leaves App
The second cut, and the seam is narrower than the first: this hook reads nothing from
App at all. Where the panels are docked, how wide they are, the header drag that moves
one between docks, the splitter drag that resizes it, and the timeline's height - none
of that touches cuts, layers, strokes or the canvas.

What stays, and why: `panelOpen`, because which panels are open is decided by the menus
and the left dock's tabs, both of which live elsewhere; and the JSX that mounts a panel
into a slot. A hook that returned the panels themselves would need every prop each one
takes, which is most of App - it would have moved the tangle rather than reduced it.

Two things the move made better rather than just relocating:

- The bottom splitter wrote `setSplitter({type:'bottom', ...})` inline in the JSX while
  panel splitters went through a helper. Both are `startPanelResize` / `startBottomResize`
  now, and the shape of a splitter object is no longer something the JSX has to know.
- The drag effect listed `[splitter]` and used PANEL_W from module scope. It takes the
  ranges as an argument now and lists them, so the dependency is real rather than
  implied - and the hook baseline is unchanged at 9, so this added no new warning.

Faithfulness, since this must change nothing. Diffed both moved blocks against the
originals:

  dock drag       one line: PANEL_IDS became the panelIds parameter
  splitter drag   the two magic numbers became named constants, PANEL_W became the
                  widthRange argument, and one statement gained braces

The validating codecs for mv_docks and mv_floats moved verbatim, including the default
float positions - a layout stored by a version that docked differently is still treated
as absent rather than restored into a shape this one cannot lay out.

App.jsx 4157 -> 4103. State 94 -> 86, effects 32 -> 31, top-level functions 153 -> 151.

Built bundle boots with no console errors. The smoke run is worth noting: it failed
first with "the API is not reachable" because I had not restarted the server - it
reports what it finds rather than what I hoped, which is the point of having it.
